sdsda

+++ lib/Screens/ItemPage/ItemPage.dart
@@ -3,9 +3,13 @@ import 'package:bhumi_app/Common/Loading/cirecularloading.dart';
 import 'package:bhumi_app/Common/Widget/Appbar.dart';
 import 'package:bhumi_app/Model/ItemDetail.dart';
 import 'package:bhumi_app/Screens/ItemPage/Imageview.dart';
+import 'package:bhumi_app/Screens/ItemPage/ItemHistory.dart';
+import 'package:bhumi_app/Screens/ItemPage/Update/ItemUpdate.dart';
 import 'package:bhumi_app/Screens/ItemPage/booking.dart';
 import 'package:bhumi_app/Service/AdditemService.dart';
+import 'package:bhumi_app/Service/Date/Dataevents.dart';
 import 'package:bhumi_app/Service/ItemService.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:photo_view/photo_view.dart';
@@ -21,11 +25,29 @@ class ItemPage extends StatefulWidget {
 
 class _ItemPageState extends State<ItemPage> {
   CalendarController _calendarController;
+  Map<DateTime,List<dynamic>> _events;
   @override
   void initState() {
     // TODO: implement initState5
     super.initState();
     _calendarController = CalendarController();
+    _events = {};
+    getdate();
+  }
+  Future<Map<DateTime,List<dynamic>>> getdate()async{
+
+    await FirebaseFirestore.instance.collection('Order').where('ProductId',isEqualTo: widget.item).get().then((value){
+      final doc =  value.docs;
+      for(DocumentSnapshot e in doc){
+        for(var i = 0; i< e.data()['BookedDates'].length ;i++){
+          DateTime convert = DateTime.parse(e.data()['BookedDates'][i]);
+          setState(() {
+            _events[convert]=[Colors.green];
+          });
+        }
+      }
+    });
+
   }
   @override
   Widget build(BuildContext context) {
@@ -49,8 +71,28 @@ class _ItemPageState extends State<ItemPage> {
         appBar: BasicAppbarwithButton(IconButton(
           color: Colors.white,
           onPressed: ()async{
-            await ItemService(itemname: widget.item).deleteItem();
-            return   Navigator.pop(context);
+
+            return showDialog(context: context,
+                builder:(BuildContext context){
+                    return AlertDialog(
+                      title: Text('Are You Sure'),
+                      content: Text('You Want To Delete ('+widget.item + ") Item ?"),
+
+                      actions: [
+                        FlatButton(
+
+                          onPressed: ()async{
+                            await ItemService(itemname: widget.item).deleteItem();
+                            Navigator.pop(context);
+                            return   Navigator.pop(context);
+                          },
+                          child: Text('Delete'),
+                        )
+                      ],
+                    );
+                }
+
+                );
 
           },
           icon: Icon(Icons.delete),
@@ -64,7 +106,7 @@ class _ItemPageState extends State<ItemPage> {
                 child: Column(
                   children: [
                     Card(
-                      elevation: 20.0,
+                      elevation: 10.0,
                       color: Colors.black,
                       child: GestureDetector(
                         onTap: (){
@@ -90,8 +132,45 @@ class _ItemPageState extends State<ItemPage> {
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
+
+
+                          RaisedButton(
+                            shape: StadiumBorder(),
+                            color: commonAssets.appbuttonColor,
+                            onPressed: (){
+                              return Navigator.push(context, PageRouteBuilder(
+                                  pageBuilder: (_,__,___) => ItemUpdate(rent: datas.rent,oldimageurl: datas.url,itemid: widget.item,productname: datas.productname,),
+                                  transitionDuration: Duration(seconds: 0)
+                              ));
+                            },
+                            child: Icon(Icons.edit,color: commonAssets.commonbuttontextcolor,),
+                          ),
+                          RaisedButton(
+                            shape: StadiumBorder(),
+                            color: commonAssets.appbuttonColor,
+                            onPressed: (){
+                              return Navigator.push(context, PageRouteBuilder(
+                                  pageBuilder: (_,__,___) => ItemHistory(productId: widget.item,),
+                                  transitionDuration: Duration(seconds: 0)
+                              ));
+                            },
+                            child: Icon(Icons.history,color: commonAssets.commonbuttontextcolor,),
+                          ),
+
+                        ],
+                      ),
+                    ),
+
+                    Center(child: Text("Rent",style: TextStyle(fontSize: 18.0,fontWeight: FontWeight.bold),)),
+                    SizedBox(height: 5,),
+                    Center(child: SelectableText(datas.rent.toString(),style: TextStyle(fontSize: 16.0,))),
+                    Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: [
                           Text("ProductName",style: TextStyle(fontSize: 18.0,fontWeight: FontWeight.bold),),
-                          Text("Rent",style: TextStyle(fontSize: 18.0,fontWeight: FontWeight.bold),),
+
                           Text("ProductId",style: TextStyle(fontSize: 18.0,fontWeight: FontWeight.bold),)
 
                         ],
@@ -103,11 +182,12 @@ class _ItemPageState extends State<ItemPage> {
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
                           Flexible(child: SelectableText(datas.productname,style: TextStyle(fontSize: 16.0,),)),
-                          Flexible(child: SelectableText(datas.rent.toString(),style: TextStyle(fontSize: 16.0,)),),
+
                           Flexible(child: SelectableText(datas.productid,style: TextStyle(fontSize: 16.0,)),)
                         ],
                       ),
                     ),
+
                     SizedBox(height: 20,),
                     Padding(
                       padding: const EdgeInsets.all(8.0),
@@ -118,6 +198,27 @@ class _ItemPageState extends State<ItemPage> {
                         borderRadius: BorderRadius.circular(10.0)
                       ),
                         child: TableCalendar(
+                          calendarStyle: CalendarStyle(
+                            selectedColor: commonAssets.selecteddates
+                          ),
+                          startingDayOfWeek: StartingDayOfWeek.monday,
+                          headerStyle: HeaderStyle(
+                            centerHeaderTitle: true,
+                            formatButtonDecoration: BoxDecoration(
+                              color: commonAssets.appbuttonColor,
+                              borderRadius: BorderRadius.circular(10.0)
+
+                            ),
+                            formatButtonTextStyle: TextStyle(
+                              color: commonAssets.commonbuttontextcolor
+                            )
+                          ),
+                          events: _events,
+                          onDaySelected: (date,events,context){
+                            setState(() {
+                              _events[_calendarController.selectedDay]=[Colors.green];
+                            });
+                          },
                           calendarController: _calendarController,
                         ),
                       ),
@@ -128,7 +229,7 @@ class _ItemPageState extends State<ItemPage> {
                       shape: StadiumBorder(),
                       onPressed: (){
                         return Navigator.push(context, PageRouteBuilder(
-                          pageBuilder: (_,__,___)=>BookingItem(itemId: datas.productid,rent: datas.rent,ordernumber: datas.ordernumber,),
+                          pageBuilder: (_,__,___)=>BookingItem(itemId: datas.productid,rent: datas.rent,events: _events,),
                           transitionDuration: Duration(seconds: 0)
                         ));
                       },
